+++ frontend/src/App.js
@@ -6,6 +6,7 @@ import StylistSignUp from './components/StylistSignUp'
 import ProfilePage from './components/ProfilePage';
 import SearchListPage from './components/SearchListPage';
 import StylistProfilePage from './components/StylistProfilePage';
+import ProfileEditPage from './components/ProfileEditPage';
 import StylistEditPage from './components/StylistEditPage';
 
 
@@ -23,7 +24,7 @@ function App() {
       <Route path='/user/:id' element={<ProfilePage />} />
       <Route path='/search/:city/:state' element={<SearchListPage />} />
       <Route path='/stylist/:id' element={<StylistProfilePage />} />
-      <Route path='/stylist/edit' element={<StylistEditPage />} />
+      <Route path='/user/edit' element={<ProfileEditPage />} />
     </Routes>
   )
 }
